requirements.txt
pyserial
pyttsx3
